requirements.txt
boto3==1.34.144
Flask==3.0.3
joblib==1.4.2
keras==3.4.1
matplotlib==3.9.1
mlflow==2.14.3
numpy==1.26.4
pandas==1.2.4
pillow==10.4.0
scikit-learn==1.5.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
urllib3==2.2.2
Werkzeug==3.0.3